moving_load/requirements.txt
numpy==1.19.3
streamlit==0.89.0
anastruct==1.2.0
matplotlib==3.2.2
Pillow==8.4.0
